// src/types/games.types.ts
export interface Platform {
  platform: {
    slug: string;
  };
}

export interface ScreenShot {
  image: string;
}

export interface Game {
  id: number;
  name: string;
  background_image: string;
  short_screenshots: ScreenShot[];
  parent_platforms: Platform[];
  rating: number;
  released: string;
  metacritic: number;
}

export interface FetchGamesResponse {
  count: number;
  results: Game[];
}